+++ src/java/com/jj/jjmod/worldgen/abstracts/WorldGenStone.java
@@ -58,27 +58,27 @@ public abstract class WorldGenStone extends WorldGenAbstract {
         int veinSize = this.getVeinSize();
 
         float f = this.rand.nextFloat() * (float) Math.PI;
-        double d0 = (double) ((float) (x + 8) + MathHelper
-                .sin(f) * (float) veinSize / 8.0F);
-        double d1 = (double) ((float) (x + 8) - MathHelper
-                .sin(f) * (float) veinSize / 8.0F);
-        double d2 = (double) ((float) (z + 8) + MathHelper
-                .cos(f) * (float) veinSize / 8.0F);
-        double d3 = (double) ((float) (z + 8) - MathHelper
-                .cos(f) * (float) veinSize / 8.0F);
-        double d4 = (double) (y + this.rand.nextInt(3) - 2);
-        double d5 = (double) (y + this.rand.nextInt(3) - 2);
+        double d0 = x + 8 + MathHelper
+                .sin(f) * veinSize / 8.0F;
+        double d1 = x + 8 - MathHelper
+                .sin(f) * veinSize / 8.0F;
+        double d2 = z + 8 + MathHelper
+                .cos(f) * veinSize / 8.0F;
+        double d3 = z + 8 - MathHelper
+                .cos(f) * veinSize / 8.0F;
+        double d4 = y + this.rand.nextInt(3) - 2;
+        double d5 = y + this.rand.nextInt(3) - 2;
 
         for (int i = 0; i < veinSize; ++i) {
             
             float f1 = (float) i / (float) veinSize;
-            double d6 = d0 + (d1 - d0) * (double) f1;
-            double d7 = d4 + (d5 - d4) * (double) f1;
-            double d8 = d2 + (d3 - d2) * (double) f1;
-            double d9 = this.rand.nextDouble() * (double) veinSize / 16.0D;
-            double d10 = (double) (MathHelper
+            double d6 = d0 + (d1 - d0) * f1;
+            double d7 = d4 + (d5 - d4) * f1;
+            double d8 = d2 + (d3 - d2) * f1;
+            double d9 = this.rand.nextDouble() * veinSize / 16.0D;
+            double d10 = (MathHelper
                     .sin((float) Math.PI * f1) + 1.0F) * d9 + 1.0D;
-            double d11 = (double) (MathHelper
+            double d11 = (MathHelper
                     .sin((float) Math.PI * f1) + 1.0F) * d9 + 1.0D;
             int j = MathHelper.floor_double(d6 - d10 / 2.0D);
             int k = MathHelper.floor_double(d7 - d11 / 2.0D);
@@ -89,19 +89,19 @@ public abstract class WorldGenStone extends WorldGenAbstract {
 
             for (int l1 = j; l1 <= i1; ++l1) {
                 
-                double d12 = ((double) l1 + 0.5D - d6) / (d10 / 2.0D);
+                double d12 = (l1 + 0.5D - d6) / (d10 / 2.0D);
 
                 if (d12 * d12 < 1.0D) {
                     
                     for (int i2 = k; i2 <= j1; ++i2) {
                         
-                        double d13 = ((double) i2 + 0.5D - d7) / (d11 / 2.0D);
+                        double d13 = (i2 + 0.5D - d7) / (d11 / 2.0D);
 
                         if (d12 * d12 + d13 * d13 < 1.0D) {
                             
                             for (int j2 = l; j2 <= k1; ++j2) {
                                 
-                                double d14 = ((double) j2 + 0.5D - d8) /
+                                double d14 = (j2 + 0.5D - d8) /
                                         (d10 / 2.0D);
 
                                 if (d12 * d12 + d13 * d13 + d14 * d14 < 1.0D) {
